% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheatds}
\alias{wheatds}
\title{Stripe rust disease severity (leaf areas infected, DS) of the wheat RIL population}
\format{
A data frame containing samples, environments, repetitions, and disease severity of the wheat RIL population.
}
\usage{
data("wheatds")
}
\description{
Stripe rust disease severity (leaf areas infected, DS) of the wheat RIL population in Yangling, Tianshui, Jiangyou.
}
\examples{
data("wheatds")
}
\keyword{datasets}
